# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 剑道馆会员管理系统 API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type MemberRole = typeof MemberRole[keyof typeof MemberRole];


export const MemberRole = {
  admin: 'admin',
  course_admin: 'course_admin',
  member: 'member',
} as const;

export type MemberStatus = typeof MemberStatus[keyof typeof MemberStatus];


export const MemberStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export interface Member {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  role: MemberRole;
  status: MemberStatus;
  joinedAt: string;
}

export interface AuthResult {
  member: Member;
}

export type MemberInputRole = typeof MemberInputRole[keyof typeof MemberInputRole];


export const MemberInputRole = {
  admin: 'admin',
  course_admin: 'course_admin',
  member: 'member',
} as const;

export interface MemberInput {
  name: string;
  email: string;
  password: string;
  phone?: string;
  role?: MemberInputRole;
}

export type MemberUpdateInputRole = typeof MemberUpdateInputRole[keyof typeof MemberUpdateInputRole];


export const MemberUpdateInputRole = {
  admin: 'admin',
  course_admin: 'course_admin',
  member: 'member',
} as const;

export interface MemberUpdateInput {
  name: string;
  email: string;
  phone?: string;
  role?: MemberUpdateInputRole;
  password?: string;
}

export interface CourseLevel {
  id: number;
  name: string;
  color: string;
  sortOrder: number;
  createdAt?: string;
}

export interface CourseLevelInput {
  name: string;
  color: string;
  sortOrder?: number;
}

export interface Announcement {
  id: number;
  title: string;
  content: string;
  isPinned: boolean;
  isPublished: boolean;
  createdAt: string;
}

export interface AnnouncementInput {
  title: string;
  content: string;
  isPinned?: boolean;
  isPublished?: boolean;
}

export interface AnnouncementUpdate {
  title?: string;
  content?: string;
  isPinned?: boolean;
  isPublished?: boolean;
}

export interface Course {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  instructor: string;
  scheduledAt: string;
  durationMinutes: number;
  maxCapacity: number;
  enrolledCount: number;
  level?: string;
}

export interface CourseInput {
  title: string;
  description?: string;
  instructor: string;
  scheduledAt: string;
  durationMinutes: number;
  maxCapacity: number;
  level?: string;
}

export interface CourseUpdate {
  title?: string;
  description?: string;
  instructor?: string;
  scheduledAt?: string;
  durationMinutes?: number;
  maxCapacity?: number;
  level?: string;
}

export type BookingStatus = typeof BookingStatus[keyof typeof BookingStatus];


export const BookingStatus = {
  confirmed: 'confirmed',
  cancelled: 'cancelled',
} as const;

export interface Booking {
  id: number;
  memberId: number;
  courseId: number;
  status: BookingStatus;
  createdAt: string;
  course?: Course;
}

export interface BookingInput {
  courseId: number;
}

export interface AdminStats {
  totalMembers: number;
  activeMembers: number;
  totalCourses: number;
  upcomingCourses: number;
  totalBookings: number;
  recentMembers: Member[];
}

export type UpdateMemberStatusBodyStatus = typeof UpdateMemberStatusBodyStatus[keyof typeof UpdateMemberStatusBodyStatus];


export const UpdateMemberStatusBodyStatus = {
  active: 'active',
  inactive: 'inactive',
} as const;

export type UpdateMemberStatusBody = {
  status: UpdateMemberStatusBodyStatus;
};