# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 剑道馆会员管理系统 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /announcements:
    get:
      operationId: listAnnouncements
      tags: [announcements]
      summary: List all announcements
      responses:
        "200":
          description: List of announcements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Announcement"
    post:
      operationId: createAnnouncement
      tags: [announcements]
      summary: Create announcement (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
  /announcements/{id}:
    patch:
      operationId: updateAnnouncement
      tags: [announcements]
      summary: Update announcement (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnouncementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Announcement"
    delete:
      operationId: deleteAnnouncement
      tags: [announcements]
      summary: Delete announcement (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all courses
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Course"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a course (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
...
  /courses/{id}:
    patch:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /course-levels:
    get:
      operationId: listCourseLevels
      tags: [course-levels]
      summary: List all course levels
      responses:
        "200":
          description: Course levels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CourseLevel"
    post:
      operationId: createCourseLevel
      tags: [course-levels]
      summary: Create a course level (course admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseLevelInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseLevel"
  /course-levels/{id}:
    patch:
      operationId: updateCourseLevel
      tags: [course-levels]
      summary: Update a course level (course admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CourseLevelInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseLevel"
    delete:
      operationId: deleteCourseLevel
      tags: [course-levels]
      summary: Delete a course level (course admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /bookings:
    get:
      operationId: listMyBookings
      tags: [bookings]
      summary: List my bookings
      responses:
        "200":
          description: My bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Book a course
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BookingInput"
      responses:
        "201":
          description: Booking created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Course full or already booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    delete:
      operationId: cancelBooking
      tags: [bookings]
      summary: Cancel a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [member]
      properties:
        member:
          $ref: "#/components/schemas/Member"
    Member:
      type: object
      required: [id, name, email, role, status, joinedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [admin, course_admin, member]
        status:
          type: string
          enum: [active, inactive]
        joinedAt:
          type: string
...
    CourseLevel:
      type: object
      required: [id, name, color, sortOrder]
      properties:
        id:
          type: integer
        name:
          type: string
        color:
          type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
    CourseLevelInput:
      type: object
      required: [name, color]
      properties:
        name:
          type: string
        color:
          type: string
        sortOrder:
          type: integer
    Announcement:
      type: object
      required: [id, title, content, isPinned, isPublished, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        isPublished:
          type: boolean
        createdAt:
          type: string
    AnnouncementInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        isPublished:
          type: boolean
    AnnouncementUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        isPinned:
          type: boolean
        isPublished:
          type: boolean
    Course:
      type: object
      required: [id, title, instructor, scheduledAt, durationMinutes, maxCapacity, enrolledCount]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        instructor:
          type: string
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        enrolledCount:
          type: integer
        level:
          type: string
    CourseInput:
      type: object
      required: [title, instructor, scheduledAt, durationMinutes, maxCapacity]
      properties:
        title:
          type: string
        description:
          type: string
        instructor:
          type: string
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        level:
          type: string
    CourseUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        instructor:
          type: string
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        maxCapacity:
          type: integer
        level:
          type: string
    Booking:
      type: object
      required: [id, memberId, courseId, status, createdAt]
      properties:
        id:
          type: integer
        memberId:
          type: integer
        courseId:
          type: integer
        status:
          type: string
          enum: [confirmed, cancelled]
        createdAt:
          type: string
        course:
          $ref: "#/components/schemas/Course"
    BookingInput:
      type: object
      required: [courseId]
      properties:
        courseId:
          type: integer
...