# lib/api-zod/src/generated/types/authResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 剑道馆会员管理系统 API
 * OpenAPI spec version: 0.1.0
 */
import type { Member } from './member';

export interface AuthResult {
  member: Member;
}