# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * 剑道馆会员管理系统 API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingStatus } from './bookingStatus';
import type { Course } from './course';

export interface Booking {
  id: number;
  memberId: number;
  courseId: number;
  status: BookingStatus;
  createdAt: string;
  course?: Course;
}